WEA-24 Move all temperature conversion logic to repository

+++ app/src/main/java/com/ayush/weatherapp/entities/DailyDataEntity.java
@@ -20,8 +20,8 @@ public class DailyDataEntity implements Parcelable {
   private String icon;
   private int sunriseTime;
   private int sunsetTime;
-  private double temperatureHigh;
-  private double temperatureLow;
+  private int temperatureHigh;
+  private int temperatureLow;
   private double windSpeed;
 
   protected DailyDataEntity(Parcel in) {
@@ -30,8 +30,8 @@ public class DailyDataEntity implements Parcelable {
     icon = in.readString();
     sunriseTime = in.readInt();
     sunsetTime = in.readInt();
-    temperatureHigh = in.readDouble();
-    temperatureLow = in.readDouble();
+    temperatureHigh = in.readInt();
+    temperatureLow = in.readInt();
     windSpeed = in.readDouble();
   }
 
@@ -78,20 +78,20 @@ public class DailyDataEntity implements Parcelable {
     this.sunsetTime = sunsetTime;
   }
 
-  public double getTemperatureHigh() {
+  public int getTemperatureHigh() {
     return temperatureHigh;
   }
 
   public void setTemperatureHigh(double temperatureHigh) {
-    this.temperatureHigh = temperatureHigh;
+    this.temperatureHigh = (int) temperatureHigh;
   }
 
-  public double getTemperatureLow() {
+  public int getTemperatureLow() {
     return temperatureLow;
   }
 
   public void setTemperatureLow(double temperatureLow) {
-    this.temperatureLow = temperatureLow;
+    this.temperatureLow = (int) temperatureLow;
   }
 
   public double getWindSpeed() {
